+++ tests/dummy/app/charts/chart/route.js
@@ -1,13 +1,12 @@
 import fetch from 'fetch';
 import Ember from 'ember';
-import ENV from '../../config/environment';
 
 export default Ember.Route.extend({
   model: function(params) {
     // TODO: default to 'bar'
     const slug = params.slug;
-    // fetch the example chart definition JSON
-    return fetch (`${ENV.rootURL}examples/${slug}.json`)
+    // fetch the example chart definition JSON from the cedar v1 docs site
+    return fetch (`http://cedar-v1.surge.sh/examples/${slug}.json`)
     .then(response => {
       // TODO: check if response is OK
       return response.json()
@@ -15,7 +14,8 @@ export default Ember.Route.extend({
         const title = `${Ember.String.capitalize(params.slug)} Example`;
         return {
           title,
-          definition
+          definition,
+          slug
         };
       });
     });