+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Fixed
 - bump cedar for better support of v0.x charts
+### Changed
+- dummy site works off of chart JSON on cedar's v1 docs site
 
 ## v1.0.0-alpha.6
 ### Fixed